Do not derive VoucherEntry from BaseObject

+++ Vouchers/Adapters/VoucherDto.cs
@@ -86,7 +86,7 @@ namespace Empiria.FinancialAccounting.Vouchers.Adapters {
       // no-op
     }
 
-    public int Id {
+    public long Id {
       get; internal set;
     }
